Move handle binding logic to academy.agent
The Agent is the only thing that uses the behavior handle binding logic
so I think it makes more sense to keep it with the Agent so the callback
complexity can be removed. This also makes it a top-level function
for easier testing.

I also found a mistake in the def of UnboundRemoteHandle.close(),
discrepancy in the read-only Handle propery attributes, and
simplified RemoteHandle.bind_to_exchange().