Rename bind_to_client and remove from RemoteHandle
Renames `bind_to_exchange` to `bind_to_client` of
`UnboundRemoteExchange` which is more accurate to what the method does:
creates a handle bound to a specific exchange client.

Removes `bind_to_client` from `RemoteHandle`, as it was not used
anywhere in the code (except for a contrived unit test), and basically
just made a replica of the current handle.