Use MP context instead of set_start_method
This caused a RuntimeError error when running:
    pytest tests
that the context had already been set. However, the error doesn't
occur when running unit and integration tests separately (as in CI):
    pytest tests/unit
    pytest tests/integration
This is because the integration tests manually call spawn_http_server
and then the unit tests call spawn_http_server once in a fixture.
Switching to use a multiprocessing context directly solves this issue of
changing the global start method after processes have already been
started.