Make logging unit tests force reconfiguration (#247)
Before this change, part of init_logging behaviour does not happen after the
first invocation: logger.basicConfig only has effect the first time it is
called.

This does affect any logging tests so far, but in an upcoming PR, I have a
new LogHandler that does not get coverage-tested in some test orderings
without this.