Add an even more extra log mode targeted at observability (#248)
extra becomes an int, a superclass of bool so existing True/False
ints continue to work as before.

It's an open question if the user-facing documentation should switch
to entirely conventional ints, rather than bools.

When a log file is configured, a second log file is created with a
format of one JSON object per line. This is easy to work with using
the `jq` command line tool and to read into Python for scripted
processing.

The new JSON logs are fairly verbose: around 1kb per entry. This is
a feature not a bug.

In the human-oriented logs, any logging extra keys beginning with
"academy." are added onto the end of the human readable log lines.

This PR introduces a noqa: SIM115 which is advice that a file
should be opened in a context manager. The logger arrangement
isn't amenable to that. Further, logging is never closed explicitly
and the log file is closed by end-of-process cleanup, so there isn't
even an explicit close.

This PR adds a single example use of extras={"academy...."} annotation,
at the start and end of an academy Manager. A separate PR will
introduce more of that.