Add logging extra keys throughout the codebase (#249)
This PR visits each log line with the basic rule:

  If someone decided they should subtitute a value into
  the log message, that should also be an extra key.

This PR does not do anything with the extra keys. The accompanying
PR #248 introduces features that will send these extra keys to log
output.

Explicit list of things this PR does not try to do:
(but could be experimented with and addressed in future)

* Add in other information even though it seems useful in passing

* Make a consistent schema/naming convention - for example,
  client_id vs agent_id vs user_id

  (and as a sub-non-goal, does not try to rename variables to be
  consistent: for example, a user_id variable that also holds agent
  IDs)

* Format the keys in any particular way: for example, agent/user IDs
  might be better formatted as full UUIDs, rather than the existing
  str/repr formats.

* Introduce code-level helpers to reduce verbosity of source code for
  creating log entries.