Replace type: ignore with a roughly equivalent assert (#278)
Depending on redis library, there is either a type error at these
lines or not:

6.4.0 ping() returns a union containing Any and so there is no type error.

7.1.0 is more strongly typed but in a way that isn't fully representable
in the type system: ping has type Awaitable[bool]|bool, but outside of the
type system it is statically true that used with an async Redis client,
ping will return the narrower type Awaitable[bool].

That means with 6.4.0 there is no type error to ignore, and so type: ignore
is itself an error, even though that same type: ignore is necessary to
typecheck with 7.1.0.

This PR fixes the type error with an assert that narrows the return
type of ping, which is valid with both of the above mentioned versions of
Redis: in 6.4.0 it narrows from Awaitable[Any]|Any to  Awaitable[Any]
and in 7.1.0 it narrows from Awaitable[bool]|bool to  Awaitable[bool].

This assert is already assumed true prior to this PR:

if ping did choose to return a bool, not an Awaitable[bool], an error would
occur one line later at the await statement.

With the static information expressed this way, there are no type errors
to ignore with either version.