Check all imports can be imported at the top level (#283)
A couple of times I've seen problems in the academy codebase to do with circular imports - most awfully, one passed CI and ended up in main, fixed by #281

The implicit assumption that #281 tries to fix is that an academy module can be imported into an environment that has not been pre-disturbed by the earlier import of other academy stuff. That's a reasonable assumption.

Given the tangled nature of academy's mutual imports, I don't think #281 will be the last circular import problem, and while I think it is worth working towards uncircularising things, the code base is where it is.

As a comparison for what warty old codebases look like, I ran the test against the parsl repo, and encountered a handful of failures: mostly Python-based configuration files that aren't imported by the main codebase and make assumptions about the environment; but one slighly-suspicious import in the visualization code that assumes that some other web server context/environment has been established before the import is made. That one is ~50% legit in my eyes.