Add error handling around bad executor specification (#298)
Prior to this PR, an arbitrary object could be passed as the executors parameter of a manager. This would result in behaviour as if no executor had been specified.

Although executors has a type specification, that is not type checked at runtime, and this is a user-facing API that can be expected to have user errors.

This PR adds a test that demonstrates the actual typo that led me here (passing an executor class, rather than invoking it).